+++ plat/rockchip/rk3399/drivers/pmu/m0_ctl.c
@@ -36,13 +36,14 @@
 #include <m0_ctl.h>
 #include <plat_private.h>
 #include <rk3399_def.h>
+#include <secure.h>
 #include <soc.h>
 
 void m0_init(void)
 {
 	/* secure config for M0 */
 	mmio_write_32(SGRF_BASE + SGRF_PMU_CON(0), WMSK_BIT(7));
-	mmio_write_32(SGRF_BASE + SGRF_SOC_CON6, WMSK_BIT(12));
+	mmio_write_32(SGRF_BASE + SGRF_SOC_CON(6), WMSK_BIT(12));
 
 	/* set the execute address for M0 */
 	mmio_write_32(SGRF_BASE + SGRF_PMU_CON(3),
